\name{gomp-package}
\alias{gomp-package}
\docType{package}
\title{
The gamma-OMP Feature Selection Algorithm
}
\description{
The \eqn{\gamma}-OMP feature selection algorithm for a wide range of response variables.
}
\details{
\tabular{lll}{
Package: \tab gomp\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2024-12-20\cr
License: \tab GPL-2\cr
}
}
\author{
Michail Tsagris \email{mtsagris@uoc.gr}.
}

\section{Maintainers}{
Michail Tsagris \email{mtsagris@uoc.gr}.
}

\references{
Tsagris M., Papadovasilakis Z., Lakiotaki K. and Tsamardinos I. (2018). Efficient feature selection on gene expression data: Which algorithm to use? BioRxiv.

Tsagris M., Papadovasilakis Z., Lakiotaki K. and Tsamardinos I. (2022). The \eqn{\gamma}-OMP algorithm for feature selection with application to gene expression data". IEEE/ACM Transactions on Computational Biology and Bioinformatics 19(2): 1214--1224.

Alharbi N. (2024). Variable selection with time-to-event data: Cox or Weibull regression? 
Communications in Statistics: Case Studies, Data Analysis and Applications (accepted for publication).
}



